feat: extract program-execution loop into public run.py
Relocate the whole-program loop (about consumption, when-block buffering,
Phase-1 exec, Phase-2 drain) and Session out of cli.run_file into a new
public run.run() that returns a structured ContractResult and performs no
I/O. cli.run_file becomes a thin wrapper: read file -> run() -> render each
result inline via the unchanged display_result (preserving amber-confirm
ordering and quiet blank-line mirroring byte-for-byte). The single when-block
buffering copy lives in run; cli._consume_when_block is a display shim over it.
